refactor: move AppPrefs to prefs package
