+++ app/src/main/java/com/osfans/trime/core/Rime.kt
@@ -4,9 +4,9 @@
 
 package com.osfans.trime.core
 
-import com.osfans.trime.data.AppPrefs
 import com.osfans.trime.data.base.DataManager
 import com.osfans.trime.data.opencc.OpenCCDictManager
+import com.osfans.trime.data.prefs.AppPrefs
 import com.osfans.trime.data.schema.SchemaManager
 import com.osfans.trime.util.appContext
 import com.osfans.trime.util.isStorageAvailable
