+++ app/src/main/java/com/osfans/trime/data/theme/ThemeManager.kt
@@ -5,8 +5,8 @@
 package com.osfans.trime.data.theme
 
 import androidx.annotation.Keep
-import com.osfans.trime.data.AppPrefs
 import com.osfans.trime.data.base.DataManager
+import com.osfans.trime.data.prefs.AppPrefs
 import com.osfans.trime.ime.symbol.TabManager
 import java.io.File
