+++ app/src/main/java/com/osfans/trime/ime/text/Candidate.kt
@@ -15,7 +15,7 @@ import android.view.MotionEvent
 import android.view.View
 import com.osfans.trime.core.CandidateListItem
 import com.osfans.trime.core.Rime
-import com.osfans.trime.data.AppPrefs
+import com.osfans.trime.data.prefs.AppPrefs
 import com.osfans.trime.data.theme.ColorManager
 import com.osfans.trime.data.theme.FontManager
 import com.osfans.trime.data.theme.ThemeManager
