+++ app/src/main/java/com/osfans/trime/ui/fragments/ProfileFragment.kt
@@ -23,8 +23,8 @@ import androidx.preference.get
 import com.osfans.trime.R
 import com.osfans.trime.core.Rime
 import com.osfans.trime.daemon.RimeDaemon
-import com.osfans.trime.data.AppPrefs
 import com.osfans.trime.data.base.DataManager
+import com.osfans.trime.data.prefs.AppPrefs
 import com.osfans.trime.ui.components.FolderPickerPreference
 import com.osfans.trime.ui.components.PaddingPreferenceFragment
 import com.osfans.trime.ui.main.MainViewModel
